refactor/seperated util functions
